refactor: split admin.ts into resource-specific endpoint files (#217)
Split the 801-line admin.ts monolith into focused files by resource type:
- admin-docs.ts: Document CRUD operations
- admin-scripts.ts: Script CRUD operations
- admin-doc-requirements.ts: Doc requirements CRUD operations
- admin-shared.ts: Shared admin authentication utility
